📖 [Docs]: Remove unused configuration options from README (#52)
## Description

This pull request updates the `README.md` file to simplify the list of
configurable settings for the action. It removes several less commonly
used settings to streamline the documentation and make it easier to
read.

Documentation cleanup:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L63-L66):
Removed the following settings from the configuration list: `Debug`,
`Verbose`, `Version`, and `Prerelease`. These settings are no longer
included in the documentation to simplify the list and focus on the most
relevant options.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] 📖 [Docs]
- [ ] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas